codebert-base.commons-lang.mid-2246.idx-19006.2.buggy

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableFloat.java
@@ -253,10 +253,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
      * @return the value associated with this instance after adding the operand
      * @since 3.5
      */
-    public float addAndGet(final Number operand) {
-        this.value += operand.floatValue();
-        return value;
-    }
+    public float addAndGet(final Number operand) { this.value += operand.shortValue(); return value; }
 
     /**
      * Increments this instance's value by {@code operand}; this method returns the value associated with the instance
